+++ src/util/vfile.h
@@ -7,10 +7,10 @@
 
 struct VFile {
 	bool (*close)(struct VFile* vf);
-	size_t (*seek)(struct VFile* vf, off_t offset, int whence);
-	size_t (*read)(struct VFile* vf, void* buffer, size_t size);
-	size_t (*readline)(struct VFile* vf, char* buffer, size_t size);
-	size_t (*write)(struct VFile* vf, void* buffer, size_t size);
+	off_t (*seek)(struct VFile* vf, off_t offset, int whence);
+	ssize_t (*read)(struct VFile* vf, void* buffer, size_t size);
+	ssize_t (*readline)(struct VFile* vf, char* buffer, size_t size);
+	ssize_t (*write)(struct VFile* vf, void* buffer, size_t size);
 	void* (*map)(struct VFile* vf, size_t size, int flags);
 	void (*unmap)(struct VFile* vf, void* memory, size_t size);
 	void (*truncate)(struct VFile* vf, size_t size);
